+++ app/signin/page.tsx
@@ -1,11 +1,11 @@
+import SignInForm from '@/components/organism/SignInPage/SignInForm';
 import React from 'react';
-import SigninForm from '@/components/organism/SignInPage/SigninForm';
 
 const page = async () => {
   return (
     <main className="no-scrollbar flex h-full w-full flex-col items-center gap-[56px] overflow-scroll px-[20px] py-[80px]">
       <h1 className="text-headline-2">알케미스트</h1>
-      <SigninForm />
+      <SignInForm />
     </main>
   );
 };
